refactor to use the word aggregate instead of hash
