+++ src/applications/simple-forms/mock-simple-forms-patterns/pages/mockArrayMultiPageAggregateItem.js
